Updating the get-raster-metadata function
Originally, the function wasn't working because it didn't account for when a raster might have an x or y value at 0, and "[U/u]nknown" isn't a recognizable input for the EML schema. Looking at the original source code, it looks like (0,0) was set to "Upper Left" as default, so I copied that into the code.

I also changed the "Unknown"s to "Lower Left" for the following set of coordinates: (x = 0, y = -)

Unknown -> Upper Left for (x = 0, y = +) & (x = +, y = 0)

Unknown -> Upper Right for (x = +, y = 0)