Function more robust
Included edits where the function doesn't error if a raster has an origin at x|y = 0, or (0,0).

Tried to use "[U/u]nknown" as an option, but the eml schema doesn't allow for that. So defaults had to be set to avoid running into that error. Should talk to Matt about the eml schema and if it can be updated or something.